make inBaseAsset as private method
